+++ tests/conftest.py
@@ -1,7 +1,6 @@
 import os
-from functools import partial
 from pathlib import Path
-from typing import Callable
+from typing import Any, Callable, List, Union
 
 import click
 import pytest
@@ -19,4 +18,13 @@ def tmp_result(tmp_path: Path) -> Path:
 @pytest.fixture
 def run_cmd() -> Callable[..., click.testing.Result]:
     runner = CliRunner()
-    return partial(runner.invoke, blackbench.main, catch_exceptions=False)
+
+    def caller(args: Union[str, List[object]] = [], **kwargs: Any) -> click.testing.Result:
+        prepped: List[str] = []
+        if isinstance(args, list):
+            prepped.extend(str(a) for a in args)
+        else:
+            prepped.append(args)
+        return runner.invoke(blackbench.main, prepped, catch_exceptions=True, **kwargs)
+
+    return caller